Adjust DEBUG_TRANS enabled code to work with current kernels and compilers
The last element in _stp_command_names is in
_stp_command_name[STP_MAX_CMD].  Thus, the array has STP_MAX_CMD+1
elements rather than STP_MAX_CMD elements.  The compiler would flag
the access beyond the end of the array due to the incorrectly sized
array.

The min operation provided by the kernel's minmax.h include does type
checking.  Needed to type cast STP_MAX_CMD to match the other argument
and avoid the compiler flagging the type mismatch as an error.

+++ runtime/transport/control.c
@@ -57,7 +57,7 @@ static ssize_t _stp_ctl_write_cmd(struct file *file, const char __user *buf, siz
 
 #if defined(DEBUG_TRANS) && (DEBUG_TRANS >= 2)
 	if (type < STP_MAX_CMD)
-		dbug_trans2("Got %s. euid=%ld, len=%d\n", _stp_command_name[min(type,STP_MAX_CMD)] ?: "?",
+		dbug_trans2("Got %s. euid=%ld, len=%d\n", _stp_command_name[min(type, (u32)STP_MAX_CMD)] ?: "?",
 			    (long)euid, (int)count);
 #endif
 
@@ -212,7 +212,7 @@ out:
 #if defined(DEBUG_TRANS) && (DEBUG_TRANS >= 2)
 	if (type < STP_MAX_CMD)
 		dbug_trans2("Completed %s (rc=%d)\n",
-                            _stp_command_name[min(type,STP_MAX_CMD)] ?: "?",
+                            _stp_command_name[min(type, (u32)STP_MAX_CMD)] ?: "?",
                             rc);
 #endif
         return rc;
